requirements.txt
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
backports.tarfile==1.2.0
billiard==4.2.1
build==1.3.0
CacheControl==0.14.3
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crashtest==0.4.1
cron_descriptor==2.0.6
diff-match-patch==20241021
distlib==0.4.0
Django==4.2.24
django-admin-rangefilter==0.13.3
django-admin-sortable2==2.2.8
django-celery-beat==2.8.1
django-environ==0.12.0
django-filter==25.1
django-import-export==4.3.9
django-jazzmin==3.0.1
django-object-actions==5.0.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
dulwich==0.24.1
exceptiongroup==1.3.0
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
kombu==5.5.4
more-itertools==10.8.0
msgpack==1.1.1
packaging==25.0
pbs-installer==2025.9.2
pkginfo==1.12.1.2
platformdirs==4.4.0
poetry-core==2.2.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
tomli==2.2.1
tomlkit==0.13.3
trove-classifiers==2025.9.11.17
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
xattr==1.2.0
zipp==3.23.0
zstandard==0.25.0
